Version19.62

+++ swad_game.h
@@ -124,7 +124,7 @@ void Gam_RequestNewMatch (void);
 
 unsigned Gam_GetNumCoursesWithGames (Hie_Level_t Scope);
 unsigned Gam_GetNumGames (Hie_Level_t Scope);
-float Gam_GetNumQstsPerCrsGame (Hie_Level_t Scope);
+double Gam_GetNumQstsPerCrsGame (Hie_Level_t Scope);
 
 void Gam_ShowTstTagsPresentInAGame (long GamCod);
